- Fixed: Backpacks didn't increase the ammo capacity to BackpackMaxAmount if   an ammo type's BackpackAmount was 0. - Fixed some minor bugs in custom skill code.
SVN r562 (trunk)

+++ src/actor.h
@@ -622,7 +622,7 @@ public:
 	fixed_t			scaleX, scaleY;		// Scaling values; FRACUNIT is normal size
 	BYTE			RenderStyle;		// Style to draw this actor with
 	DWORD			renderflags;		// Different rendering flags
-	WORD			picnum;				// Draw this instead of sprite if != 0xffff
+	int				picnum;				// Draw this instead of sprite if != 0xffff
 	SWORD			TIDtoHate;			// TID of things to hate (0 if none)
 	DWORD			effects;			// [RH] see p_effect.h
 	fixed_t			alpha;
